+++ src/Symfony/Bundle/FrameworkBundle/Tests/Templating/GlobalVariablesTest.php
@@ -26,18 +26,6 @@ class GlobalVariablesTest extends TestCase
         $this->globals = new GlobalVariables($this->container);
     }
 
-    /**
-     * @group legacy
-     */
-    public function testLegacyGetSecurity()
-    {
-        $securityContext = $this->getMock('Symfony\Component\Security\Core\SecurityContextInterface');
-
-        $this->assertNull($this->globals->getSecurity());
-        $this->container->set('security.context', $securityContext);
-        $this->assertSame($securityContext, $this->globals->getSecurity());
-    }
-
     public function testGetUserNoTokenStorage()
     {
         $this->assertNull($this->globals->getUser());